Added the grammar structure classes
I added some of the classes that are needed for the grammar system that
I am implementing. I added the word-classes of the English language and
I am soon gonna add a recognition engine for the different classes.
This is how the human(s) are gonna be able to learn stuff. The grammar
listener is for looking through sentences and more. Sentences takes one
parameter which is a sentence with a period and a uppercase letter to
start off with. The human class now also inherits from the Creature
class. The Creature class has the information about a living Creature.
I also fixed some bugs within the Brain capacity generator.