Added a new disease
Fixed a lot of bugs in different classes all over the engine. I also
added a born-system for borning a baby from a human, but of course the
human must be a female for it to work. I also fixed some bugs in the
LifeListener, but it's disabled for now - because of more bugs.